refactor(BankAccountStatus): Change format of BankAccountStatus Enum
